+++ core/src/jsMain/kotlin/dev/fritz2/dom/html/svgtag.kt
@@ -2,7 +2,6 @@
 
 package dev.fritz2.dom.html
 
-import dev.fritz2.dom.HtmlTag
 import kotlinx.browser.document
 import kotlinx.coroutines.Job
 import org.w3c.dom.Element
@@ -11,7 +10,7 @@ import org.w3c.dom.svg.SVGElement
 const val SVG_XMLNS = "http://www.w3.org/2000/svg"
 
 /**
- * Exposes the JavaScript [SVGElement](https://developer.mozilla.org/en-US/docs/Web/API/SVGElement) to Kotlin
+ * Implementation of [Tag] to represent the JavaScript [SVGElement](https://developer.mozilla.org/en-US/docs/Web/API/SVGElement) to Kotlin
  */
 class SvgTag<out E : Element>(tagName: String, id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
     HtmlTag<SVGElement>(tagName, id, baseClass, job, scope) {
